pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "stripe>=12.0.0",
    "eth-account>=0.13.6",
    "eth-utils>=5.2.0",
    "oauthlib>=3.2.2",
    "wtforms>=3.2.1",
    "trafilatura>=2.0.0",
    "web3>=7.10.0",
    "eth-abi>=5.2.0",
    "python-dotenv>=1.1.0",
    "openai>=1.74.0",
    "svglib>=1.5.1",
]
